+++ packages/nhost-js/CHANGELOG.md
@@ -1,3 +1,7 @@
+# dev
+
+- feat (packages/nhost-js): add decoded token to session storage #37
+
 # 5.0.0-beta.7
 
 - fix (packages/nhost-js): extract query name from TypedDocumentNode #36